PEMA-107 Do not use phony targets as prerequisites
If we do that, the rule will always run. Instead we need to use real files as prerequisites.

Reference: https://www.gnu.org/software/make/manual/make.html#Phony-Targets:
"A phony target should not be a prerequisite of a real target file; if it is, its recipe will be run every time make goes to update that file."