Updated android build script
Makefiles for android need to point towards backend, not old sequoia with all that implies, such as changing .mk, makefiles and includes.

Added libCxx11 to the task list with the same logic as libPEpTransport as they work the same.

Added header link to identity_api.cc this is purely for the sake of android studio and not required, but android studio complains about missing headers without it.

Updated Engine_version_code, will move it to a gradle variable later.
Todo: 
Investigate why JNI dislikes using identity without a global variable during handshakes.